Move Agent HTTP header initialization to process startup
